+++ kleidicv/src/resize/resize_linear_api.cpp
@@ -76,6 +76,9 @@ kleidicv_error_t kleidicv_resize_linear_stripe_u8(
     const uint8_t *src, size_t src_stride, size_t src_width, size_t src_height,
     size_t y_begin, size_t y_end, uint8_t *dst, size_t dst_stride,
     size_t dst_width, size_t dst_height, size_t channels) {
+  CHECK_POINTER_AND_STRIDE(src, src_stride, src_height);
+  CHECK_POINTER_AND_STRIDE(dst, dst_stride, dst_height);
+
   if (src_width == 0 || src_height == 0) {
     return KLEIDICV_OK;
   }